+++ packages/core/src/model/BaseModel.ts
@@ -85,7 +85,7 @@ interface IBaseModel {
   moveText(deltaX: number, deltaY: number): void;
   updateText(value: string): void;
   setSelected(flag: boolean): void;
-  setZIndex(zindex?: number): void;
+  setZIndex(zIndex?: number): void;
   /**
    * 设置Node|Edge等model的状态
    * @param state 状态