+++ packages/core/src/model/node/BaseNodeModel.ts
@@ -297,7 +297,7 @@ export default class BaseNodeModel implements IBaseNodeModel {
    */
   isAllowConnectedAsSource(
     target: BaseNodeModel,
-    soureAnchor: AnchorConfig,
+    sourceAnchor: AnchorConfig,
     targetAnchor: AnchorConfig,
   ): ConnectRuleResult | Boolean {
     const rules = !this.hasSetSourceRules
@@ -308,7 +308,7 @@ export default class BaseNodeModel implements IBaseNodeModel {
     let msg: string;
     for (let i = 0; i < rules.length; i++) {
       const rule = rules[i];
-      if (!rule.validate.call(this, this, target, soureAnchor, targetAnchor)) {
+      if (!rule.validate.call(this, this, target, sourceAnchor, targetAnchor)) {
         isAllPass = false;
         msg = rule.message;
         break;
@@ -331,7 +331,7 @@ export default class BaseNodeModel implements IBaseNodeModel {
    */
   isAllowConnectedAsTarget(
     source: BaseNodeModel,
-    soureAnchor: AnchorConfig,
+    sourceAnchor: AnchorConfig,
     targetAnchor: AnchorConfig,
   ): ConnectRuleResult | Boolean {
     const rules = !this.hasSetTargetRules
@@ -342,7 +342,7 @@ export default class BaseNodeModel implements IBaseNodeModel {
     let msg: string;
     for (let i = 0; i < rules.length; i++) {
       const rule = rules[i];
-      if (!rule.validate.call(this, source, this, soureAnchor, targetAnchor)) {
+      if (!rule.validate.call(this, source, this, sourceAnchor, targetAnchor)) {
         isAllPass = false;
         msg = rule.message;
         break;
@@ -632,8 +632,8 @@ export default class BaseNodeModel implements IBaseNodeModel {
   }
 
   @action
-  setZIndex(zindex = 1): void {
-    this.zIndex = zindex;
+  setZIndex(zIndex = 1): void {
+    this.zIndex = zIndex;
   }
 
   @action