+++ packages/extension/examples/custom-edge/custom-edge.js
@@ -15,22 +15,22 @@ class CustomEdgeModel extends PolylineEdgeModel {
     if (currentPositionList.length > 1) {
       let [ x1, y1 ] = currentPositionList[0].split(',');
       let [ x2, y2 ] = currentPositionList[1].split(',');
-      let distence = 50;
+      let distance = 50;
       x1 = Number(x1)
       y1 = Number(y1)
       x2 = Number(x2)
       y2 = Number(y2)
       if (x1 === x2) { // 垂直
         if (y2 < y1) {
-          distence = -50;
+          distance = -50;
         }
-        position.y = y1 + distence;
+        position.y = y1 + distance;
         position.x = x1;
       } else {
         if (x2 < x1) {
-          distence = -50;
+          distance = -50;
         }
-        position.x = x1 + distence;
+        position.x = x1 + distance;
         position.y = y1 - 10;
       }
     }